requirements.txt
Django
djangorestframework
psycopg2
django-filter

